Fix watchOS arm64_32 debug build https://bugs.webkit.org/show_bug.cgi?id=301267 rdar://163179870
Unreviewed build fix.

OSStatus is long on non-64bit.
NSUInteger is unsigned int on non-64bit.
Long is at least size of int, so fix by casting to long.

* Source/WebCore/platform/network/mac/UTIUtilities.mm:
(WebCore::setImageSourceAllowableTypes):
* Source/WebKit/NetworkProcess/cocoa/NetworkSessionCocoa.mm:
(WebKit::NetworkSessionCocoa::initializeNSURLSessionsInSet):
* Source/WebKit/Shared/Cocoa/CoreIPCError.mm:
(WebKit::CoreIPCError::CoreIPCError):

Canonical link: https://commits.webkit.org/302019@main

+++ Source/WebCore/platform/network/mac/UTIUtilities.mm
@@ -156,7 +156,7 @@ void setImageSourceAllowableTypes(const Vector<String>& supportedImageTypes)
     std::call_once(onceFlag, [supportedImageTypes] {
         auto allowableTypes = createNSArray(supportedImageTypes);
         auto status = CGImageSourceSetAllowableTypes((__bridge CFArrayRef)allowableTypes.get());
-        RELEASE_ASSERT_WITH_MESSAGE(supportedImageTypes.isEmpty() || status == noErr, "CGImageSourceSetAllowableTypes() returned error: %d.", status);
+        RELEASE_ASSERT_WITH_MESSAGE(supportedImageTypes.isEmpty() || status == noErr, "CGImageSourceSetAllowableTypes() returned error: %ld.", static_cast<long>(status));
     });
 }
